+++ src/main/java/com/pezfa/inventario/models/VentaDetalle.java
@@ -8,7 +8,7 @@ package com.pezfa.inventario.models;
 public class VentaDetalle 
 {
     private Venta venta;
-    private int cantidad;
+    private float cantidad;
     public VentaDetalle()
     {
         
@@ -24,12 +24,12 @@ public class VentaDetalle
         this.venta = venta;
     }
 
-    public int getCantidad()
+    public float getCantidad()
     {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad)
+    public void setCantidad(float cantidad)
     {
         this.cantidad = cantidad;
     }
